put the db configuration in the server.js file, setup express
